Adding test cases checkbox/environemnts list components
