+++ src/resources/js/pages/sessions/register/info.vue
@@ -59,210 +59,11 @@
                             <h3 class="card-title">Select use cases</h3>
                         </div>
                         <div class="card-body pt-0 pl-0">
-                            <ul
-                                class="list-group overflow-auto"
+                            <test-case-checkboxes
                                 style="max-height: 320px;"
-                            >
-                                <li
-                                    class="list-group-item"
-                                    v-for="useCase in useCases.data"
-                                >
-                                    <div class="d-flex align-items-center">
-                                        <b
-                                            class="dropdown-toggle"
-                                            v-b-toggle="
-                                                `use-case-${useCase.id}`
-                                            "
-                                        >
-                                            {{ useCase.name }}
-                                        </b>
-                                        <button
-                                            type="button"
-                                            class="btn btn-link py-0 font-weight-normal text-decoration-none"
-                                            @click.prevent="toggleCheckboxes"
-                                        >
-                                            <icon name="checkbox" />
-                                        </button>
-                                    </div>
-
-                                    <b-collapse
-                                        :id="`use-case-${useCase.id}`"
-                                        visible
-                                    >
-                                        <ul
-                                            class="list-group"
-                                            v-if="
-                                                collect(useCase.testCases)
-                                                    .where(
-                                                        'behavior',
-                                                        'positive'
-                                                    )
-                                                    .count()
-                                            "
-                                        >
-                                            <li
-                                                class="list-group-item border-0 py-0"
-                                            >
-                                                <div
-                                                    class="d-flex align-items-center"
-                                                >
-                                                    <span
-                                                        class="d-inline-block dropdown-toggle py-2 font-weight-medium"
-                                                        v-b-toggle="
-                                                            `positive-test-cases-${useCase.id}`
-                                                        "
-                                                    >
-                                                        Happy flow
-                                                    </span>
-                                                    <button
-                                                        type="button"
-                                                        class="btn btn-link py-0 font-weight-normal text-decoration-none"
-                                                        @click.prevent="
-                                                            toggleCheckboxes
-                                                        "
-                                                    >
-                                                        <icon name="checkbox" />
-                                                    </button>
-                                                </div>
-
-                                                <b-collapse
-                                                    :id="`positive-test-cases-${useCase.id}`"
-                                                    visible
-                                                >
-                                                    <ul class="list-group">
-                                                        <li
-                                                            class="list-group-item"
-                                                            v-for="testCase in collect(
-                                                                useCase.testCases
-                                                            )
-                                                                .where(
-                                                                    'behavior',
-                                                                    'positive'
-                                                                )
-                                                                .all()"
-                                                        >
-                                                            <label
-                                                                class="form-check mb-0"
-                                                            >
-                                                                <input
-                                                                    :value="
-                                                                        testCase.id
-                                                                    "
-                                                                    v-model="
-                                                                        form.test_cases
-                                                                    "
-                                                                    type="checkbox"
-                                                                    class="form-check-input"
-                                                                />
-                                                                <span
-                                                                    class="form-check-label"
-                                                                >
-                                                                    {{
-                                                                        testCase.name
-                                                                    }}
-                                                                    <icon
-                                                                        name="lock"
-                                                                        v-if="
-                                                                            !testCase.public
-                                                                        "
-                                                                        class="text-muted"
-                                                                    />
-                                                                </span>
-                                                            </label>
-                                                        </li>
-                                                    </ul>
-                                                </b-collapse>
-                                            </li>
-                                        </ul>
-
-                                        <ul
-                                            class="list-group"
-                                            v-if="
-                                                collect(useCase.testCases)
-                                                    .where(
-                                                        'behavior',
-                                                        'negative'
-                                                    )
-                                                    .count()
-                                            "
-                                        >
-                                            <li
-                                                class="list-group-item border-0 py-0"
-                                            >
-                                                <div
-                                                    class="d-flex align-items-center"
-                                                >
-                                                    <span
-                                                        class="d-inline-block dropdown-toggle py-2 font-weight-medium"
-                                                        v-b-toggle="
-                                                            `negative-test-cases-${useCase.id}`
-                                                        "
-                                                    >
-                                                        Unhappy flow
-                                                    </span>
-                                                    <button
-                                                        type="button"
-                                                        class="btn btn-link py-0 font-weight-normal text-decoration-none"
-                                                        @click.prevent="
-                                                            toggleCheckboxes
-                                                        "
-                                                    >
-                                                        <icon name="checkbox" />
-                                                    </button>
-                                                </div>
-
-                                                <b-collapse
-                                                    :id="`negative-test-cases-${useCase.id}`"
-                                                    visible
-                                                >
-                                                    <ul class="list-group">
-                                                        <li
-                                                            class="list-group-item"
-                                                            v-for="testCase in collect(
-                                                                useCase.testCases
-                                                            )
-                                                                .where(
-                                                                    'behavior',
-                                                                    'negative'
-                                                                )
-                                                                .all()"
-                                                        >
-                                                            <label
-                                                                class="form-check mb-0"
-                                                            >
-                                                                <input
-                                                                    :value="
-                                                                        testCase.id
-                                                                    "
-                                                                    v-model="
-                                                                        form.test_cases
-                                                                    "
-                                                                    type="checkbox"
-                                                                    class="form-check-input"
-                                                                />
-                                                                <span
-                                                                    class="form-check-label"
-                                                                >
-                                                                    {{
-                                                                        testCase.name
-                                                                    }}
-                                                                    <icon
-                                                                        name="lock"
-                                                                        v-if="
-                                                                            !testCase.public
-                                                                        "
-                                                                        class="text-muted"
-                                                                    />
-                                                                </span>
-                                                            </label>
-                                                        </li>
-                                                    </ul>
-                                                </b-collapse>
-                                            </li>
-                                        </ul>
-                                    </b-collapse>
-                                </li>
-                            </ul>
+                                :useCases="useCases"
+                                v-model="form.test_cases"
+                            />
                             <div
                                 class="text-danger small mt-3"
                                 v-if="$page.errors.test_cases"
@@ -294,10 +95,12 @@
 
 <script>
 import Layout from '@/layouts/sessions/register';
+import TestCaseCheckboxes from '@/components/sessions/test-case-checkboxes';
 
 export default {
     components: {
         Layout,
+        TestCaseCheckboxes,
     },
     props: {
         session: {
@@ -328,21 +131,6 @@ export default {
         };
     },
     methods: {
-        toggleCheckboxes(e) {
-            const btn = e.target;
-            const closestParentList = btn.closest('.list-group-item');
-            const checkboxes = Array.from(
-                closestParentList.querySelectorAll('input[type="checkbox"]')
-            );
-            const isChecked = checkboxes.every(
-                (checkbox) => checkbox.checked === true
-            );
-
-            checkboxes.forEach((checkbox) => {
-                checkbox.checked = !isChecked;
-                checkbox.dispatchEvent(new Event('change'));
-            });
-        },
         submit() {
             this.sending = true;
             this.$inertia
